Move internal functions into base class/file.
First step in breaking up the monolith and cleaning up the __init__.py
pollution.
